value(&mut self, i: i32)

+++ rust/src/encodings.rs
@@ -51,5 +51,5 @@ pub trait Decoder {
 
     fn take(&mut self, indices: &Int32Array) -> Result<Box<dyn Array>>;
 
-    fn value(&mut self, i: usize) -> Result<Box<dyn Scalar>>;
+    fn value(&mut self, i: i32) -> Result<Box<dyn Scalar>>;
 }
